Functionality implemented 1. Aliyun uploader; 2. Project plan implemented; 3. Project management completed.

+++ zzyl-nursing-platform/src/main/java/com/zzyl/nursing/mapper/NursingPlanMapper.java
@@ -1,15 +1,14 @@
 package com.zzyl.nursing.mapper;
 
+import java.util.List;
 import com.baomidou.mybatisplus.core.mapper.BaseMapper;
 import org.apache.ibatis.annotations.Mapper;
-
-import java.util.List;
 import com.zzyl.nursing.domain.NursingPlan;
 
 /**
  * 护理计划Mapper接口
  * 
- * @author jasperhu
+ * @author LaoYe
  * @date 2024-11-09
  */
 @Mapper
@@ -21,7 +20,7 @@ public interface NursingPlanMapper extends BaseMapper<NursingPlan>
      * @param id 护理计划主键
      * @return 护理计划
      */
-    public NursingPlan selectNursingPlanById(Long id);
+    public NursingPlan selectNursingPlanById(Integer id);
 
     /**
      * 查询护理计划列表
@@ -53,7 +52,7 @@ public interface NursingPlanMapper extends BaseMapper<NursingPlan>
      * @param id 护理计划主键
      * @return 结果
      */
-    public int deleteNursingPlanById(Long id);
+    public int deleteNursingPlanById(Integer id);
 
     /**
      * 批量删除护理计划
@@ -61,5 +60,5 @@ public interface NursingPlanMapper extends BaseMapper<NursingPlan>
      * @param ids 需要删除的数据主键集合
      * @return 结果
      */
-    public int deleteNursingPlanByIds(Long[] ids);
+    public int deleteNursingPlanByIds(Integer[] ids);
 }
